refactor: JWT package refactoring
