+++ src/main/java/yagu/yagu/common/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package yagu.yagu.common.security;
+package yagu.yagu.common.jwt;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -8,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
+import yagu.yagu.common.jwt.JwtTokenProvider;
 
 import java.io.IOException;
