+++ src/main/java/yagu/yagu/common/jwt/JwtTokenProvider.java
@@ -1,7 +1,5 @@
-package yagu.yagu.common.security;
+package yagu.yagu.common.jwt;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
@@ -10,9 +8,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Component;
+import yagu.yagu.common.jwt.JwtConfig;
 
 import java.security.Key;
 import java.util.Date;
