+++ src/main/java/yagu/yagu/common/security/SecurityConfig.java
@@ -16,6 +16,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import yagu.yagu.common.jwt.JwtAuthenticationFilter;
+import yagu.yagu.common.jwt.JwtTokenProvider;
 import yagu.yagu.user.service.AuthService;
 
 import java.io.IOException;
